Size is not grandeur, and territory does not make a nation mandatory

+++ 0x00-hello_world/6-size.c
@@ -14,7 +14,7 @@ int main(void)
 	float f;
 
 	printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(d));
-	printf("Size of an int: %d byte(s)\n", (int) sizeof(a));
+	printf("Size of an int: %u byte(s)\n", (unsigned long) sizeof(a));
 	printf("Size of long int: %ld byte(s)\n", (long int) sizeof(b));
 	printf("Size of a long long int: %lld byte(s)\n", (long long int) sizeof(c));
 	printf("Size of a float: %lu byte(s)\n", (unsigned long) sizeof(f));
